New changes: Category added:
The category variable was added to the __init__ method of the GlobalPlugin class to define a category for the add-on.
The chooseGesture method was implemented to allow changing the keyboard shortcut in the "Input Gestures" dialog.

Script category assigned:

The @script decorator was extended with the category="Progress Reader" parameter to assign the script to the new category.
With these changes, the user can modify the keyboard shortcut for the add-on in the "Input Gestures" dialog, and the script will be listed under the "Progress Reader" category.